combine SupportedMimeTypes and MediaType into one
