+++ app/jobs/upgrade/backfill_image_derivatives.rb
@@ -2,7 +2,7 @@ class Upgrade::BackfillImageDerivatives < Upgrade::BackfillDerivativesBase
   queue_as :low
 
   def mime_types
-    SupportedMimeTypes.image_types.without(Mime[:svg])
+    MediaType.image_types.without(Mime[:svg])
   end
 
   def derivative
