+++ app/lib/file_handlers/image_tag.rb
@@ -1,6 +1,6 @@
 class FileHandlers::ImageTag < FileHandlers::Base
   ENVIRONMENTS = [:browser, :preview_frame].freeze
-  INPUT_TYPES = SupportedMimeTypes.image_types
+  INPUT_TYPES = MediaType.image_types
 
   def self.component
     Components::Renderers::ImageTag
