+++ app/lib/file_handlers/video_tag.rb
@@ -1,6 +1,6 @@
 class FileHandlers::VideoTag < FileHandlers::Base
   ENVIRONMENTS = [:browser, :preview_frame].freeze
-  INPUT_TYPES = SupportedMimeTypes.video_types.freeze
+  INPUT_TYPES = MediaType.video_types.freeze
 
   def self.component
     Components::Renderers::VideoTag
