+++ app/models/model_file.rb
@@ -61,19 +61,19 @@ class ModelFile < ApplicationRecord
   end
 
   def is_image?
-    SupportedMimeTypes.image_extensions.include? extension
+    MediaType.image_extensions.include? extension
   end
 
   def is_video?
-    SupportedMimeTypes.video_extensions.include? extension
+    MediaType.video_extensions.include? extension
   end
 
   def is_document?
-    SupportedMimeTypes.document_extensions.include? extension
+    MediaType.document_extensions.include? extension
   end
 
   def is_3d_model?
-    SupportedMimeTypes.model_extensions.include? extension
+    MediaType.model_extensions.include? extension
   end
 
   def has_render?
