+++ app/models/site_settings.rb
@@ -5,7 +5,7 @@ class SiteSettings < RailsSettings::Base
   field :model_tags_filter_stop_words, type: :boolean, default: true
   field :model_tags_tag_model_directory_name, type: :boolean, default: false
   field :model_tags_stop_words_locale, type: :string, default: "en"
-  field :model_tags_custom_stop_words, type: :array, default: SupportedMimeTypes.indexable_extensions
+  field :model_tags_custom_stop_words, type: :array, default: MediaType.indexable_extensions
   field :model_tags_auto_tag_new, type: :string, default: "!new"
   field :model_ignored_files, type: :array, default: [
     /^\.[^.]+/, # Hidden files starting with .
